# gen/external/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package externalapi

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn1AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn4AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"POST": "Content-Type,X-Bootstrap-Token",
	}
	rn6AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
	}
	rn8AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn10AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PUT":    "Authorization,Content-Type",
	}
	rn11AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn13AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PUT":    "Authorization,Content-Type",
	}
	rn14AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn16AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn17AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn19AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PUT":    "Authorization,Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "auth/"

				if l := len("auth/"); len(elem) >= l && elem[0:l] == "auth/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'm': // Prefix: "me"

					if l := len("me"); len(elem) >= l && elem[0:l] == "me" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleAuthMeGetRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 't': // Prefix: "tokens"

					if l := len("tokens"); len(elem) >= l && elem[0:l] == "tokens" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleAuthTokensListRequest([0]string{}, elemIsEscaped, w, r)
						case "POST":
							s.handleAuthTokensCreateRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET,POST",
								allowedHeaders: rn4AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'b': // Prefix: "bootstrap"
							origElem := elem
							if l := len("bootstrap"); len(elem) >= l && elem[0:l] == "bootstrap" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "POST":
									s.handleAuthTokensBootstrapRequest([0]string{}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "POST",
										allowedHeaders: rn3AllowedHeaders,
										acceptPost:     "application/json",
										acceptPatch:    "",
									})
								}

								return
							}

							elem = origElem
						}
						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleAuthTokensDeleteRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "DELETE",
									allowedHeaders: rn6AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				}

			case 'b': // Prefix: "broadcasts"

				if l := len("broadcasts"); len(elem) >= l && elem[0:l] == "broadcasts" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleBroadcastsListRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleBroadcastsCreateRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET,POST",
							allowedHeaders: rn8AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleBroadcastsDeleteRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleBroadcastsGetRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PUT":
							s.handleBroadcastsUpdateRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,GET,PUT",
								allowedHeaders: rn10AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'c': // Prefix: "contacts"

				if l := len("contacts"); len(elem) >= l && elem[0:l] == "contacts" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleContactsListRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleContactsCreateRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET,POST",
							allowedHeaders: rn11AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleContactsDeleteRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleContactsGetRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PUT":
							s.handleContactsUpdateRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,GET,PUT",
								allowedHeaders: rn13AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'e': // Prefix: "event"

				if l := len("event"); len(elem) >= l && elem[0:l] == "event" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case '-': // Prefix: "-actions"

					if l := len("-actions"); len(elem) >= l && elem[0:l] == "-actions" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleEventActionsListRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn14AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 's': // Prefix: "s"

					if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleEventsCreateRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn16AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 's': // Prefix: "segments"

				if l := len("segments"); len(elem) >= l && elem[0:l] == "segments" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleSegmentsListRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleSegmentsCreateRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET,POST",
							allowedHeaders: rn17AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleSegmentsDeleteRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleSegmentsGetRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PUT":
							s.handleSegmentsUpdateRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,GET,PUT",
								allowedHeaders: rn19AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "auth/"

				if l := len("auth/"); len(elem) >= l && elem[0:l] == "auth/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'm': // Prefix: "me"

					if l := len("me"); len(elem) >= l && elem[0:l] == "me" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = AuthMeGetOperation
							r.summary = ""
							r.operationID = "Auth_meGet"
							r.operationGroup = ""
							r.pathPattern = "/auth/me"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 't': // Prefix: "tokens"

					if l := len("tokens"); len(elem) >= l && elem[0:l] == "tokens" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = AuthTokensListOperation
							r.summary = ""
							r.operationID = "Auth_tokensList"
							r.operationGroup = ""
							r.pathPattern = "/auth/tokens"
							r.args = args
							r.count = 0
							return r, true
						case "POST":
							r.name = AuthTokensCreateOperation
							r.summary = ""
							r.operationID = "Auth_tokensCreate"
							r.operationGroup = ""
							r.pathPattern = "/auth/tokens"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'b': // Prefix: "bootstrap"
							origElem := elem
							if l := len("bootstrap"); len(elem) >= l && elem[0:l] == "bootstrap" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "POST":
									r.name = AuthTokensBootstrapOperation
									r.summary = ""
									r.operationID = "Auth_tokensBootstrap"
									r.operationGroup = ""
									r.pathPattern = "/auth/tokens/bootstrap"
									r.args = args
									r.count = 0
									return r, true
								default:
									return
								}
							}

							elem = origElem
						}
						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "DELETE":
								r.name = AuthTokensDeleteOperation
								r.summary = ""
								r.operationID = "Auth_tokensDelete"
								r.operationGroup = ""
								r.pathPattern = "/auth/tokens/{id}"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			case 'b': // Prefix: "broadcasts"

				if l := len("broadcasts"); len(elem) >= l && elem[0:l] == "broadcasts" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = BroadcastsListOperation
						r.summary = ""
						r.operationID = "Broadcasts_list"
						r.operationGroup = ""
						r.pathPattern = "/broadcasts"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = BroadcastsCreateOperation
						r.summary = ""
						r.operationID = "Broadcasts_create"
						r.operationGroup = ""
						r.pathPattern = "/broadcasts"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = BroadcastsDeleteOperation
							r.summary = ""
							r.operationID = "Broadcasts_delete"
							r.operationGroup = ""
							r.pathPattern = "/broadcasts/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = BroadcastsGetOperation
							r.summary = ""
							r.operationID = "Broadcasts_get"
							r.operationGroup = ""
							r.pathPattern = "/broadcasts/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PUT":
							r.name = BroadcastsUpdateOperation
							r.summary = ""
							r.operationID = "Broadcasts_update"
							r.operationGroup = ""
							r.pathPattern = "/broadcasts/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			case 'c': // Prefix: "contacts"

				if l := len("contacts"); len(elem) >= l && elem[0:l] == "contacts" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ContactsListOperation
						r.summary = ""
						r.operationID = "Contacts_list"
						r.operationGroup = ""
						r.pathPattern = "/contacts"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = ContactsCreateOperation
						r.summary = ""
						r.operationID = "Contacts_create"
						r.operationGroup = ""
						r.pathPattern = "/contacts"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = ContactsDeleteOperation
							r.summary = ""
							r.operationID = "Contacts_delete"
							r.operationGroup = ""
							r.pathPattern = "/contacts/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = ContactsGetOperation
							r.summary = ""
							r.operationID = "Contacts_get"
							r.operationGroup = ""
							r.pathPattern = "/contacts/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PUT":
							r.name = ContactsUpdateOperation
							r.summary = ""
							r.operationID = "Contacts_update"
							r.operationGroup = ""
							r.pathPattern = "/contacts/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			case 'e': // Prefix: "event"

				if l := len("event"); len(elem) >= l && elem[0:l] == "event" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case '-': // Prefix: "-actions"

					if l := len("-actions"); len(elem) >= l && elem[0:l] == "-actions" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = EventActionsListOperation
							r.summary = ""
							r.operationID = "EventActions_list"
							r.operationGroup = ""
							r.pathPattern = "/event-actions"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 's': // Prefix: "s"

					if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = EventsCreateOperation
							r.summary = ""
							r.operationID = "Events_create"
							r.operationGroup = ""
							r.pathPattern = "/events"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 's': // Prefix: "segments"

				if l := len("segments"); len(elem) >= l && elem[0:l] == "segments" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = SegmentsListOperation
						r.summary = ""
						r.operationID = "Segments_list"
						r.operationGroup = ""
						r.pathPattern = "/segments"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = SegmentsCreateOperation
						r.summary = ""
						r.operationID = "Segments_create"
						r.operationGroup = ""
						r.pathPattern = "/segments"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = SegmentsDeleteOperation
							r.summary = ""
							r.operationID = "Segments_delete"
							r.operationGroup = ""
							r.pathPattern = "/segments/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = SegmentsGetOperation
							r.summary = ""
							r.operationID = "Segments_get"
							r.operationGroup = ""
							r.pathPattern = "/segments/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PUT":
							r.name = SegmentsUpdateOperation
							r.summary = ""
							r.operationID = "Segments_update"
							r.operationGroup = ""
							r.pathPattern = "/segments/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			}

		}
	}
	return r, false
}